+++ sonoff/xdrv_interface.ino
@@ -225,7 +225,7 @@ bool XdrvCall(uint8_t Function)
 {
   bool result = false;
 
-  for (uint32_t x = 0; x < xdrv_present; x++) {
+  for (uint8_t x = 0; x < xdrv_present; x++) {
 //    WifiAddDelayWhenDisconnected();
     result = xdrv_func_ptr[x](Function);
